+++ workspaces/tables/src/components/OuterWrapper.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const OuterWrapper: React.FC = ({ children }) => {
+  return <div className={'relative h-full w-full'}>{children}</div>;
+};
+
+export default OuterWrapper;