+++ packages/select/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @heathmont/moon-select
 
+## 9.10.0
+
+### Minor Changes
+
+- Add Table component
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.10.0
+  - @heathmont/moon-themes@9.10.0
+  - @heathmont/moon-utils@9.10.0
+
 ## 9.9.0
 
 ### Minor Changes