+++ workspaces/themes/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-themes-tw
 
+## 9.10.0
+
+### Minor Changes
+
+- Add Table component
+
 ## 9.9.0
 
 ### Minor Changes